Fix qrbox border placement (#555)
* Add display: block; to the video element

A video element behave like an inline element, which mean it honours
whitespace around. When in a div with whitespace it adds empty spaces.

ref: https://stackoverflow.com/a/36397260

* Update logic to display the qrbox border elements

Using the bottom allow the 4 boxes on the bottom of the qrbox to be
pixel perfect on the bottom of the qrbox area, instead of relying on the
qrbox size that can be with a floating point.